Clean up the implementation, as this seems the simplest solution
